+++ src/Storages/MergeTree/MergeTreeData.cpp
@@ -5028,10 +5028,6 @@ void MergeTreeData::swapActivePart(MergeTreeData::DataPartPtr part_copy, DataPar
             ssize_t diff_rows = part_copy->rows_count - original_active_part->rows_count;
             increaseDataVolume(diff_bytes, diff_rows, /* parts= */ 0);
 
-            /// Move parts are non replicated operations, so we take lock here.
-            /// All other locks are taken in StorageReplicatedMergeTree
-            lockSharedData(*part_copy, /* replace_existing_lock */ true);
-
             return;
         }
     }
@@ -7980,7 +7976,7 @@ PartitionCommandsResultInfo MergeTreeData::freezePartitionsByMatcher(
         {
             // Store metadata for replicated table.
             // Do nothing for non-replicated.
-            createAndStoreFreezeMetadata(disk, part, fs::path(backup_part_path) / part->getDataPartStorage().getPartDirectory());
+            createAndStoreFreezeMetadata(disk, part, fs::path(backup_part_path) / part->getDataPartStorage().getPartDirectory()); // TODO leave as virtual
         };
 
         IDataPartStorage::ClonePartParams params
