+++ src/Storages/MergeTree/MutateFromLogEntryTask.cpp
@@ -130,91 +130,22 @@ ReplicatedMergeMutateTaskBase::PrepareResult MutateFromLogEntryTask::prepare()
     reserved_space = StorageReplicatedMergeTree::reserveSpace(estimated_space_for_result, source_part->getDataPartStorage());
     future_mutated_part->updatePath(storage, reserved_space.get());
 
+    // Created prepareAfterReservation because we cannot get IDisk object from source_part
+    // to acquire zero-copy lock
+    // TODO: make zero-copy part derived class
+    if (!prepareAfterReservation(source_part_name, source_part, estimated_space_for_result, need_to_check_missing_part_in_fetch))
+        return PrepareResult{
+            .prepared_successfully = false,
+            .need_to_check_missing_part_in_fetch = need_to_check_missing_part_in_fetch,
+            .part_log_writer = part_log_writer,
+        };
+
+
     table_lock_holder = storage.lockForShare(
             RWLockImpl::NO_QUERY, (*storage_settings_ptr)[MergeTreeSetting::lock_acquire_timeout_for_background_operations]);
     StorageMetadataPtr metadata_snapshot = storage.getInMemoryMetadataPtr();
 
     transaction_ptr = std::make_unique<MergeTreeData::Transaction>(storage, NO_TRANSACTION_RAW);
-
-    if ((*storage_settings_ptr)[MergeTreeSetting::allow_remote_fs_zero_copy_replication])
-    {
-        if (auto disk = reserved_space->getDisk(); disk->supportZeroCopyReplication())
-        {
-            if (storage.findReplicaHavingCoveringPart(entry.new_part_name, true))
-            {
-                LOG_DEBUG(log, "Mutation of part {} finished by some other replica, will download mutated part", entry.new_part_name);
-                return PrepareResult{
-                    .prepared_successfully = false,
-                    .need_to_check_missing_part_in_fetch = true,
-                    .part_log_writer = part_log_writer,
-                };
-            }
-
-            if ((*storage_settings_ptr)[MergeTreeSetting::zero_copy_merge_mutation_min_parts_size_sleep_before_lock] != 0 &&
-                estimated_space_for_result >= (*storage_settings_ptr)[MergeTreeSetting::zero_copy_merge_mutation_min_parts_size_sleep_before_lock])
-            {
-                /// In zero copy replication only one replica execute merge/mutation, others just download merged parts metadata.
-                /// Here we are trying to metigate the skew of merges execution because of faster/slower replicas.
-                /// Replicas can be slow because of different reasons like bigger latency for ZooKeeper or just slight step behind because of bigger queue.
-                /// In this case faster replica can pick up all merges execution, especially large merges while other replicas can just idle. And even in this case
-                /// the fast replica is not overloaded because amount of executing merges don't affect the ability to aquite locks for new merges.
-                ///
-                /// So here we trying to solve it with the simplest solution -- sleep random time up to 500ms for 1GB part and up to 7 seconds for 300GB part.
-                /// It can sound too much, but we are trying to acquire these locks in background tasks which can be scheduled each 5 seconds or so.
-                double start_to_sleep_seconds = std::logf((*storage_settings_ptr)[MergeTreeSetting::zero_copy_merge_mutation_min_parts_size_sleep_before_lock].value);
-                uint64_t right_border_to_sleep_ms = static_cast<uint64_t>((std::log(estimated_space_for_result) - start_to_sleep_seconds + 0.5) * 1000);
-                uint64_t time_to_sleep_milliseconds = std::min<uint64_t>(10000UL, std::uniform_int_distribution<uint64_t>(1, 1 + right_border_to_sleep_ms)(rng));
-
-                LOG_INFO(log, "Mutation size is {} bytes (it's more than sleep threshold {}) so will intentionally sleep for {} ms to allow other replicas to took this big mutation",
-                    estimated_space_for_result, (*storage_settings_ptr)[MergeTreeSetting::zero_copy_merge_mutation_min_parts_size_sleep_before_lock], time_to_sleep_milliseconds);
-
-                std::this_thread::sleep_for(std::chrono::milliseconds(time_to_sleep_milliseconds));
-            }
-
-            zero_copy_lock = storage.tryCreateZeroCopyExclusiveLock(entry.new_part_name, disk);
-
-            if (!zero_copy_lock || !zero_copy_lock->isLocked())
-            {
-                LOG_DEBUG(
-                    log,
-                    "Mutation of part {} started by some other replica, will wait for it and mutated merged part. Number of tries {}",
-                    entry.new_part_name,
-                    entry.num_tries);
-                storage.watchZeroCopyLock(entry.new_part_name, disk);
-
-                return PrepareResult{
-                    .prepared_successfully = false,
-                    .need_to_check_missing_part_in_fetch = false,
-                    .part_log_writer = part_log_writer,
-                };
-            }
-            if (storage.findReplicaHavingCoveringPart(entry.new_part_name, /* active */ false))
-            {
-                /// Why this if still needed? We can check for part in zookeeper, don't find it and sleep for any amount of time. During this sleep part will be actually committed from other replica
-                /// and exclusive zero copy lock will be released. We will take the lock and execute mutation one more time, while it was possible just to download the part from other replica.
-                ///
-                /// It's also possible just because reads in [Zoo]Keeper are not lineariazable.
-                ///
-                /// NOTE: In case of mutation and hardlinks it can even lead to extremely rare dataloss (we will produce new part with the same hardlinks, don't fetch the same from other replica), so this check is important.
-                ///
-                /// In case of DROP_RANGE on fast replica and stale replica we can have some failed select queries in case of zero copy replication.
-                zero_copy_lock->lock->unlock();
-
-                LOG_DEBUG(
-                    log,
-                    "We took zero copy lock, but mutation of part {} finished by some other replica, will release lock and download "
-                    "mutated part to avoid data duplication",
-                    entry.new_part_name);
-                return PrepareResult{
-                    .prepared_successfully = false,
-                    .need_to_check_missing_part_in_fetch = true,
-                    .part_log_writer = part_log_writer,
-                };
-            }
-
-            LOG_DEBUG(log, "Zero copy lock taken, will mutate part {}", entry.new_part_name);
-        }
-    }
 
     task_context = Context::createCopy(storage.getContext());
     task_context->makeQueryContextForMutate(*storage.getSettings());
@@ -288,11 +219,12 @@ bool MutateFromLogEntryTask::finalize(ReplicatedMergeMutateTaskBase::PartLogWrit
         throw;
     }
 
-    if (zero_copy_lock)
-    {
-        LOG_DEBUG(log, "Removing zero-copy lock");
-        zero_copy_lock->lock->unlock();
-    }
+    // Move to MutateZeroCopyFromLogEntryTask
+    // if (zero_copy_lock)
+    // {
+    //     LOG_DEBUG(log, "Removing zero-copy lock");
+    //     zero_copy_lock->lock->unlock();
+    // }
 
     /** With `ZSESSIONEXPIRED` or `ZOPERATIONTIMEOUT`, we can inadvertently roll back local changes to the parts.
          * This is not a problem, because in this case the entry will remain in the queue, and we will try again.
@@ -305,5 +237,102 @@ bool MutateFromLogEntryTask::finalize(ReplicatedMergeMutateTaskBase::PartLogWrit
     return true;
 }
 
+
+bool MutateZeroCopyFromLogEntryTask::finalize(ReplicatedMergeMutateTaskBase::PartLogWriter write_part_log)
+{
+    MutateFromLogEntryTask::finalize(write_part_log);
+
+    if (zero_copy_lock)
+    {
+        LOG_DEBUG(log, "Removing zero-copy lock");
+        zero_copy_lock->lock->unlock();
+    }
+}
+
+bool MutateZeroCopyFromLogEntryTask::prepareAfterReservation(
+        const String & source_part_name,
+        MergeTreeData::DataPartPtr active_containing_part,
+        size_t estimated_space_for_result,
+        bool & need_to_check_missing_part_in_fetch)
+{
+    if (!(*storage_settings_ptr)[MergeTreeSetting::allow_remote_fs_zero_copy_replication])
+        return true;
+    
+    if (auto disk = reserved_space->getDisk(); disk->supportZeroCopyReplication())
+    {
+        if (storage.findReplicaHavingCoveringPart(entry.new_part_name, true))
+        {
+            LOG_DEBUG(log, "Mutation of part {} finished by some other replica, will download mutated part", entry.new_part_name);
+            need_to_check_missing_part_in_fetch = true;
+            return false;
+        }
+
+        if ((*storage_settings_ptr)[MergeTreeSetting::zero_copy_merge_mutation_min_parts_size_sleep_before_lock] != 0 &&
+            estimated_space_for_result >= (*storage_settings_ptr)[MergeTreeSetting::zero_copy_merge_mutation_min_parts_size_sleep_before_lock])
+        {
+            /// In zero copy replication only one replica execute merge/mutation, others just download merged parts metadata.
+            /// Here we are trying to metigate the skew of merges execution because of faster/slower replicas.
+            /// Replicas can be slow because of different reasons like bigger latency for ZooKeeper or just slight step behind because of bigger queue.
+            /// In this case faster replica can pick up all merges execution, especially large merges while other replicas can just idle. And even in this case
+            /// the fast replica is not overloaded because amount of executing merges don't affect the ability to aquite locks for new merges.
+            ///
+            /// So here we trying to solve it with the simplest solution -- sleep random time up to 500ms for 1GB part and up to 7 seconds for 300GB part.
+            /// It can sound too much, but we are trying to acquire these locks in background tasks which can be scheduled each 5 seconds or so.
+            double start_to_sleep_seconds = std::logf((*storage_settings_ptr)[MergeTreeSetting::zero_copy_merge_mutation_min_parts_size_sleep_before_lock].value);
+            uint64_t right_border_to_sleep_ms = static_cast<uint64_t>((std::log(estimated_space_for_result) - start_to_sleep_seconds + 0.5) * 1000);
+            uint64_t time_to_sleep_milliseconds = std::min<uint64_t>(10000UL, std::uniform_int_distribution<uint64_t>(1, 1 + right_border_to_sleep_ms)(rng));
+
+            LOG_INFO(log, "Mutation size is {} bytes (it's more than sleep threshold {}) so will intentionally sleep for {} ms to allow other replicas to took this big mutation",
+                estimated_space_for_result, (*storage_settings_ptr)[MergeTreeSetting::zero_copy_merge_mutation_min_parts_size_sleep_before_lock], time_to_sleep_milliseconds);
+
+            std::this_thread::sleep_for(std::chrono::milliseconds(time_to_sleep_milliseconds));
+        }
+
+        zero_copy_lock = storage.tryCreateZeroCopyExclusiveLock(entry.new_part_name, disk);
+
+        if (!zero_copy_lock || !zero_copy_lock->isLocked())
+        {
+            LOG_DEBUG(
+                log,
+                "Mutation of part {} started by some other replica, will wait for it and mutated merged part. Number of tries {}",
+                entry.new_part_name,
+                entry.num_tries);
+            storage.watchZeroCopyLock(entry.new_part_name, disk);
+
+            return PrepareResult{
+                .prepared_successfully = false,
+                .need_to_check_missing_part_in_fetch = false,
+                .part_log_writer = part_log_writer,
+            };
+        }
+        if (storage.findReplicaHavingCoveringPart(entry.new_part_name, /* active */ false))
+        {
+            /// Why this if still needed? We can check for part in zookeeper, don't find it and sleep for any amount of time. During this sleep part will be actually committed from other replica
+            /// and exclusive zero copy lock will be released. We will take the lock and execute mutation one more time, while it was possible just to download the part from other replica.
+            ///
+            /// It's also possible just because reads in [Zoo]Keeper are not lineariazable.
+            ///
+            /// NOTE: In case of mutation and hardlinks it can even lead to extremely rare dataloss (we will produce new part with the same hardlinks, don't fetch the same from other replica), so this check is important.
+            ///
+            /// In case of DROP_RANGE on fast replica and stale replica we can have some failed select queries in case of zero copy replication.
+            zero_copy_lock->lock->unlock();
+
+            LOG_DEBUG(
+                log,
+                "We took zero copy lock, but mutation of part {} finished by some other replica, will release lock and download "
+                "mutated part to avoid data duplication",
+                entry.new_part_name);
+            return PrepareResult{
+                .prepared_successfully = false,
+                .need_to_check_missing_part_in_fetch = true,
+                .part_log_writer = part_log_writer,
+            };
+        }
+
+        LOG_DEBUG(log, "Zero copy lock taken, will mutate part {}", entry.new_part_name);
+    }
+    
+    return true;
+}
 
 }
